+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Bug fixes
 
 - Fixed expansion of undefined environment variables (now consistent with command line behaviour, where such args are empty strings)
+- Fixed a bug in inferring extensions of Dictionary and Array types
 
 ## 0.17.0